requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Pygments==2.13.0
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
psycopg2-binary==2.9.3
gunicorn==20.1.0
pytest==7.1.3
pytest-django==4.5.2
celery==5.2.7
redis==4.3.4